Extract helpers from existing tests
